Use HDF5 from AnalysisBase
This commit removes dependence on the LCG versions of HDF5 and
replaces it with the version in recent AnalysisBase builds. Since the
AnalysisBase version is more recent, I also had to make a few other
adjustments.

I also removed the old multi-step variable copying code, which meant
we could remove around 500 lines of code.